Sprite + Fix of AI + new debug feature

+++ Common/src/main/java/gsbs/common/components/Hitbox.java
@@ -2,9 +2,9 @@ package gsbs.common.components;
 
 public class Hitbox extends Component{
 
-    private float width, height;
-    private float x, y;
-    private float hitboxPadding;
+    private double width, height;
+    private double x, y;
+    private double hitboxPadding;
 
     public Hitbox(float width, float height, float x, float y) {
         hitboxPadding = Math.min(width, height) * 0.35f;
@@ -15,10 +15,10 @@ public class Hitbox extends Component{
     }
 
     public void set(float x, float y){
-        float spriteWidth = this.width * 2.0f;
-        float spriteHeight = this.height * 2.0f;
-        float offsetX = (spriteWidth - (width + hitboxPadding)) / 2.0f;
-        float offsetY = (spriteHeight - (height + hitboxPadding)) / 2.0f;
+        double spriteWidth = this.width * 2.0f;
+        double spriteHeight = this.height * 2.0f;
+        double offsetX = (spriteWidth - (width + hitboxPadding)) / 2.0f;
+        double offsetY = (spriteHeight - (height + hitboxPadding)) / 2.0f;
         this.x = x + offsetX;
         this.y = y + offsetY;
     }
@@ -36,19 +36,19 @@ public class Hitbox extends Component{
         return right && left && bottom && top;
     }
 
-    public float getWidth(){
+    public double getWidth(){
         return width;
     }
 
-    public float getHeight(){
+    public double getHeight(){
         return height;
     }
 
-    public float getX(){
+    public double getX(){
         return x;
     }
 
-    public float getY(){
+    public double getY(){
         return y;
     }
 }
